[2021-09-24] 4.10-11 - 4.10-11 완료 - App.jsx 에서 function app 을 App component 로 변경 - Navbar 추가
