+++ src/App.css
@@ -1,3 +1,5 @@
+/** @format */
+
 * {
   box-sizing: border-box;
 }
@@ -6,6 +8,30 @@ button {
   outline: none;
   border: 0;
   cursor: pointer;
+}
+
+.navbar {
+  font-size: 2rem;
+  padding: 1em;
+  background-color: blanchedalmond;
+}
+
+.navbar-logo {
+  color: green;
+  margin-right: 0.5em;
+}
+
+.navbar-count {
+  display: inline-block;
+  text-align: center;
+  font-size: 2rem;
+  width: 2.5rem;
+  height: 2.5rem;
+  line-height: 2.5rem;
+  margin-left: 0.2em;
+  background-color: green;
+  color: white;
+  border-radius: 50%;
 }
 
 .habit-li {
@@ -34,14 +60,15 @@ button {
   background-color: transparent;
 }
 
-.habit-button:hover{
+.habit-button:hover {
   opacity: 0.8;
 }
 
-.habit-increase, .habit-decrease {
+.habit-increase,
+.habit-decrease {
   color: goldenrod;
 }
 
 .habit-delete {
-  color: red
+  color: red;
 }
