+++ src/component/habit.jsx
@@ -16,7 +16,6 @@ class Habit extends Component {
   };
 
   render() {
-    console.log(this.props);
     const { name, count } = this.props.habit;
 
     return (
