+++ src/component/habits.jsx
@@ -4,44 +4,16 @@ import React, { Component } from 'react';
 import Habit from './habit';
 
 class Habits extends Component {
-  state = {
-    habits: [
-      { id: 1, name: 'Reading', count: 0 },
-      { id: 2, name: 'Running', count: 0 },
-      { id: 3, name: 'Coding', count: 0 }
-    ]
-  };
-
-  handleHabitIncrease = (habit) => {
-    const habits = [...this.state.habits];
-    const index = habits.indexOf(habit);
-    habits[index].count++;
-    this.setState({ habits });
-  };
-
-  handleHabitDecrease = (habit) => {
-    const habits = [...this.state.habits];
-    const index = habits.indexOf(habit);
-    const count = habits[index].count - 1;
-    habits[index].count = count < 0 ? 0 : count;
-    this.setState({ habits });
-  };
-
-  handleHabitDelete = (habit) => {
-    const habits = this.state.habits.filter((item) => item.id !== habit.id);
-    this.setState({ habits });
-  };
-
   render() {
     return (
       <ul>
-        {this.state.habits.map((habit) => (
+        {this.props.habits.map((habit) => (
           <Habit
             key={habit.id}
             habit={habit}
-            onIncrement={this.handleHabitIncrease}
-            onDecrement={this.handleHabitDecrease}
-            onDelete={this.handleHabitDelete}
+            onIncrement={this.props.onIncrement}
+            onDecrement={this.props.onDecrement}
+            onDelete={this.props.onDelete}
           />
         ))}
       </ul>
