+++ src/component/navbar.jsx
@@ -1,0 +1,17 @@
+/** @format */
+
+import React, { Component } from 'react';
+
+class Navbar extends Component {
+  render() {
+    return (
+      <div className="navbar">
+        <i className="navbar-logo fas fa-leaf"></i>
+        <span>Habit Tracker</span>
+        <span className="navbar-count">{this.props.habitCount}</span>
+      </div>
+    );
+  }
+}
+
+export default Navbar;
